Move resolveSegmentAssignments to Audience Panel
