+++ components/AudiencePanel.tsx
@@ -5,7 +5,42 @@ import React, { useMemo } from 'react'
 import LabelValuePanel from '@/components/LabelValuePanel'
 import SegmentsTable from '@/components/SegmentsTable'
 import VariationsTable from '@/components/VariationsTable'
-import { ExperimentFull, Segment, SegmentType } from '@/models'
+import { ExperimentFull, Segment, SegmentAssignment, SegmentType } from '@/models'
+
+/**
+ * Resolves the segment ID of the segment assignment with the actual segment.
+ * If the ID cannot be resolved, then an `Error` will be thrown.
+ *
+ * @param
+ * @param segments - The segments to associate with the assignments.
+ * @throws {Error} When unable to resolve a segment ID with one of the supplied
+ *   segments.
+ */
+function resolveSegmentAssignments(
+  segmentAssignments: SegmentAssignment[],
+  segments: Segment[],
+): {
+  segment: Segment
+  isExcluded: boolean
+}[] {
+  const segmentsById: { [segmentId: string]: Segment } = {}
+  segments.forEach((segment) => (segmentsById[segment.segmentId] = segment))
+
+  return segmentAssignments.map((segmentAssignment) => {
+    const segment = segmentsById[segmentAssignment.segmentId]
+
+    if (!segment) {
+      throw Error(
+        `Failed to lookup segment with ID ${segmentAssignment.segmentId} for assignment with ID ${segmentAssignment.segmentAssignmentId}.`,
+      )
+    }
+
+    return {
+      segment,
+      isExcluded: segmentAssignment.isExcluded,
+    }
+  })
+}
 
 /**
  * Renders the audience information of an experiment in a panel component.
@@ -16,8 +51,8 @@ import { ExperimentFull, Segment, SegmentType } from '@/models'
  */
 function AudiencePanel({ experiment, segments }: { experiment: ExperimentFull; segments: Segment[] }) {
   const segmentsByType = useMemo(
-    () => _.groupBy(experiment.resolveSegmentAssignments(segments), _.property('segment.type')),
-    [experiment, segments],
+    () => _.groupBy(resolveSegmentAssignments(experiment.segmentAssignments, segments), _.property('segment.type')),
+    [experiment.segmentAssignments, segments],
   )
 
   const countryResolvedSegmentAssignments = segmentsByType[SegmentType.Country] ?? []
